update tests and model.save
